+++ settings/src/main/java/com/team23/settings/domain/repositories/AchievementRepository.kt
@@ -1,0 +1,5 @@
+package com.team23.settings.domain.repositories
+
+interface AchievementRepository {
+    suspend fun resetAchievements()
+}