+++ settings/src/main/java/com/team23/settings/domain/repositories/FactRepository.kt
@@ -1,0 +1,5 @@
+package com.team23.settings.domain.repositories
+
+interface FactRepository {
+    suspend fun resetFacts()
+}