+++ settings/src/main/java/com/team23/settings/presentation/viewmodels/SettingsVM.kt
@@ -4,13 +4,19 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.CreationExtras
+import com.team23.settings.domain.usecases.ResetDatabaseUseCase
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 class SettingsVM @AssistedInject constructor(
-    @Assisted val changeStatusAndNavigationColors: () -> Unit,
+    @Assisted("changeStatusAndNavigationColors") val changeStatusAndNavigationColors: () -> Unit,
+    @Assisted("resetAchievementData") val resetAchievementData: () -> Unit,
+    private val resetDatabaseUseCase: ResetDatabaseUseCase,
 ) : ViewModel() {
     /**
      * True: Forced dark mode
@@ -25,17 +31,23 @@ class SettingsVM @AssistedInject constructor(
         changeStatusAndNavigationColors()
     }
 
-    fun onLanguageChanged(value : Int) {
+    fun onLanguageChanged(value: Int) {
         // TODO
     }
 
     fun onResetData() {
-
+        resetAchievementData()
+        viewModelScope.launch(Dispatchers.IO) {
+            resetDatabaseUseCase()
+        }
     }
 
     @AssistedFactory
     interface Factory {
-        fun create(changeStatusAndNavigationColors: () -> Unit): SettingsVM
+        fun create(
+            @Assisted("changeStatusAndNavigationColors") changeStatusAndNavigationColors: () -> Unit,
+            @Assisted("resetAchievementData") resetAchievementData: () -> Unit
+        ): SettingsVM
     }
 
     @Suppress("UNCHECKED_CAST")
@@ -43,9 +55,13 @@ class SettingsVM @AssistedInject constructor(
         fun provideFactory(
             assistedFactory: Factory,
             changeStatusAndNavigationColors: () -> Unit,
+            resetAchievementData: () -> Unit,
         ): ViewModelProvider.Factory = object : ViewModelProvider.Factory {
             override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
-                return assistedFactory.create(changeStatusAndNavigationColors) as T
+                return assistedFactory.create(
+                    changeStatusAndNavigationColors,
+                    resetAchievementData
+                ) as T
             }
         }
     }